+++ mslicer/src/app/mod.rs
@@ -29,7 +29,7 @@ use crate::{
     windows::{self, Tab},
 };
 use common::{progress::CombinedProgress, units::Milimeter};
-use slicer::{format::FormatSliceFile, slicer::Slicer};
+use slicer::slicer::Slicer;
 
 pub mod config;
 pub mod history;
@@ -178,16 +178,15 @@ impl App {
             ],
             move || {
                 let slice_operation = slice_operation.as_ref().unwrap();
-                let slice_result = slicer.slice_format();
-                let layers = slice_result.layers();
+                let (mut file, voxels) = slicer.slice();
+                let layers = file.info().layers as usize;
 
-                let slice_config = slice_result.slice_config();
-                let volume = (slice_result.voxels() as f32 * slice_config.voxel_volume()).convert();
+                let slice_config = slicer.slice_config();
+                let volume = (voxels as f32 * slice_config.voxel_volume()).convert();
                 let print_time = slice_config.print_time(layers as u32);
 
-                let preview_image = slice_operation.preview_image();
-                let mut file = FormatSliceFile::from_slice_result(&preview_image, slice_result);
                 post_processing.process(&mut file, post_process);
+                file.set_preview(&slice_operation.preview_image());
 
                 slice_operation.add_result(SliceResult {
                     file: Arc::new(file),