+++ mslicer/src/app/project/mod.rs
@@ -1,11 +1,11 @@
 use std::path::PathBuf;
 
 use crate::app::project::model::Model;
-use common::{progress::CombinedProgress, slice::SliceConfig};
-use slicer::{
-    format::FormatSliceFile,
-    post_process::{anti_alias::AntiAlias, elephant_foot_fixer::ElephantFootFixer},
+use common::{
+    progress::CombinedProgress,
+    slice::{DynSlicedFile, SliceConfig},
 };
+use slicer::post_process::{anti_alias::AntiAlias, elephant_foot_fixer::ElephantFootFixer};
 
 pub mod model;
 pub mod storage;
@@ -34,7 +34,7 @@ impl Project {
 }
 
 impl PostProcessing {
-    pub fn process(&self, file: &mut FormatSliceFile, progress: CombinedProgress<2>) {
+    pub fn process(&self, file: &mut DynSlicedFile, progress: CombinedProgress<2>) {
         self.elephant_foot_fixer
             .post_slice(file, progress[0].clone());
         self.anti_alias.post_slice(file, progress[1].clone());