+++ mslicer/src/windows/slice_operation.rs
@@ -33,7 +33,7 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
         if let Some(completion) = slice_operation.completion() {
             let mut result = app.slice_operation.as_ref().unwrap().result();
             let result = result.as_mut().unwrap();
-            let format = result.file.as_format();
+            let format = result.file.format();
 
             ui.horizontal(|ui| {
                 ui.label(format!("Slicing completed in {completion}!"));