+++ slicer/src/lib.rs
@@ -1,7 +1,6 @@
 //! This module contains the definition for the [`mesh::Mesh`] struct along with methods for the actual slicing operations.
 
 pub mod builder;
-pub mod format;
 pub mod geometry;
 pub mod half_edge;
 pub mod mesh;