+++ slicer/src/post_process/anti_alias.rs
@@ -1,14 +1,13 @@
 use common::{
     progress::Progress,
     serde::{Deserializer, Serializer},
+    slice::{DynSlicedFile, SliceLayerIterator},
 };
 use libblur::{
     AnisotropicRadius, BlurImageMut, EdgeMode, EdgeMode2D, FastBlurChannels, ThreadingPolicy,
 };
 use rayon::iter::{ParallelBridge, ParallelIterator};
 use serde::{Deserialize, Serialize};
-
-use crate::format::FormatSliceFile;
 
 #[derive(Clone, Serialize, Deserialize)]
 pub struct AntiAlias {
@@ -17,25 +16,27 @@ pub struct AntiAlias {
 }
 
 impl AntiAlias {
-    pub fn post_slice(&self, file: &mut FormatSliceFile, progress: Progress) {
+    pub fn post_slice(&self, file: &mut DynSlicedFile, progress: Progress) {
         if !self.enabled {
             return;
         }
 
         progress.set_total(file.info().layers as u64);
-        file.iter_mut_layers().par_bridge().for_each(|mut layer| {
-            progress.add_complete(1);
-            let (width, height) = (layer.width(), layer.height());
-            let mut image =
-                BlurImageMut::borrow(&mut layer, width, height, FastBlurChannels::Plane);
-            libblur::fast_gaussian_next(
-                &mut image,
-                AnisotropicRadius::new(self.radius as u32),
-                ThreadingPolicy::Single,
-                EdgeMode2D::new(EdgeMode::Clamp),
-            )
-            .unwrap();
-        });
+        SliceLayerIterator::new(file)
+            .par_bridge()
+            .for_each(|mut layer| {
+                progress.add_complete(1);
+                let (width, height) = (layer.size.x as u32, layer.size.y as u32);
+                let mut image =
+                    BlurImageMut::borrow(layer.inner_mut(), width, height, FastBlurChannels::Plane);
+                libblur::fast_gaussian_next(
+                    &mut image,
+                    AnisotropicRadius::new(self.radius as u32),
+                    ThreadingPolicy::Single,
+                    EdgeMode2D::new(EdgeMode::Clamp),
+                )
+                .unwrap();
+            });
 
         progress.set_finished();
     }